Set bindings directly when they are edited
This is done using the new stripes.setBindings function (see STRPCORE-3).

Unfortunately, the change in the Redux store that this triggers causes
a re-render, which defocuses the text-area and loses the cursor
position with it.

To prevent this, following a suggestion from
https://github.com/gcanti/tcomb-form/issues/51#issuecomment-70993827
we wrap the real component in a wrapper component whose only purpose
is to have the shouldComponentUpdate lifecycle method return
false.

This makes things work, but it's ugly. It may be that using redux-form
would solve this for us; or tnere may be a better solution futher down
the (long) discussion that suggested the present approach.

Sort of fixes STRPCORE-3.