Change translation strings
1. Use German instead of French, as the current development VM knows
about the former as a valid locale but not the latter.

2. Provide different strings: ones which are translated on rendering
by React components rather than directly as strings. We need to do
this for now, as we have no way to translate outside the context of
React-rendered HTML: see STCOR-52.

These modified strings allow us to go ahead with UIORG-24.