separate params from path. Refs UIORG-69
A resource's path and params values should be handled separately so as
to avoid problems URL encoding, joining multiple parameters, etc. It's
possible to just tack the params onto the path and hope that no other
operation comes along and tries to amend the query string, but this is
exploiting a loophole, not good design.

Hattip @skomorokh.