make stripes-core a peerDependency
stripes-core MUST be a peerDependency in order to ensure that only a
single copy of it exists in the bundle. Multiple copies mean the
provider and consumer of objects via the context API can have different
versions, which does not work. A provider cannot register a consumer
created with a different version.