prevent resource overwrite by EntryManager
When EntryManager was refactored into EntryWrapper and EntryManager to
get URL resources working, that had the side effect of overwriting the
resources property if resources were passed in via props. To be fair,
passing in resources withouth renaming the prop was really stupid. Now,
it's less stupid, so it works.