validation use of locations before removal
Check whether a location is in use by a holdings or item record before
attempting to remove it. This allows us to be more confident that a
DELETE request will be successful if we make it that far, and allows us
to show a nicer, more informative error message if we know such a
request would fail.

Refs UIORG-86