provide default isActive value to new locations
It appears that under some circumstances redux-form will only submit
dirty form fields if a field was not provided to `props.initialValues`.
This is particularly relevant when creating new items because it means
that new item will not be correctly initialized if, say, the form has a
dropdown menu with a default selection.

https://github.com/erikras/redux-form/issues/701 may be relevant, at
least as far as understanding the behavior.

This behavior has never been observed locally but did appear reliably on
folio-testing. What a mystery.

Fixes UIORG-109