wait longer before attempting delete
Logging during testing shows that if the script immediately starts
looping through the list of institutions after they load, it gets a list
that contains some institutions and some libraries. This makes no sense
whatsoever, and does not reflect what is displayed in the UI, and yet
that's what it reports on the console. If we wait, the UI and console
appear to be in sync.

I give up. I don't want to wait. I don't understand why we're out of
sync, or even _how_ we can be out of sync in this manner. Maybe if we
had the wrong list entirely, i.e. a list of all libraries, that would
make sense. But a list of half libraries and half institutions? I got
nothin' to explain that.