UITEN-34: Enable translation of headings on Settings
The solution here is to provide translation strings for every plugin,
and dynamically generate IDs  for them that can be passed to FormattedMessage.
the downside of this, of course, is that when we add new plugins, we have to
remember to add new translation strings here, or they will just show up as
IDs in the settings panel.  I can't see any better way to do it, though.

https://issues.folio.org/browse/UITEN-34