UITEN-128 do not warn about dangling references (#174)
In short, this PR removes checks for dangling references from
course-reserves, an optional Okapi module. It does so because, as that
module is optional and this one is required, this one sits below it in
the hierarchy of modules and therefore cannot know about the interfaces,
and more importantly, the permissions required to access the endpoints
defined by those interfaces. Thus, even though the queries were
protected correctly by `stripes.ifInterface(...)` conditions, the
queries would still fail because there was no way for the
location-managment pset (`ui-tenant-settings.settings.location`) to
include those permissions.

I feel bad removing this code -- which, to be clear, does its job
perfectly cromulently -- but I have to because there is no way (at least
at present) for it to run with the proper permissions.

Refs UITEN-128